feat: extract calculateNextReview to utility
